Add css for positiong of the .card
This consists of creating a .card-container to use flex on all the .card classes. We use flex-wrap in order to make the .card wrap to new row if there is not enough space in that row to house X amount of .card.

Then, in .card, we set flex-basis to 300px to give it the hypothetical width to start with AND make flex-grow and shrink 0. That way, when the container resizes, it will not resize the .card elements but instead push them using wrap.