Add missing code from solution
Two very important missing pieces:
(1) Did not add flex: 1 to .content, and (2) did not display flex body element.

The result of not doing (1) was that the .content would not grow to eat up any available space if there was. This would cause the footer to float up

(2) the issue was that even if i did (1), without display: flex on the body i.e. the parent container of .content, .content won't grow to push the footer down. If i want flex to work on items parents need to be flex containers.